refactor(daemon): move cluster-metrics to metrics/ to break the shim↔k8s cycle
cluster-metrics was the only shim→k8s import while k8s imports shim heavily.
Moving it to the neutral metrics/ directory removes the directory-level cycle.
No re-export stub is left behind; all seven importers updated.
