+++ packages/daemon/src/k8s/driver.ts
@@ -1,4 +1,4 @@
-import { LaunchTimer, noopClusterMetrics, type ClusterMetrics } from './cluster-metrics.js'
+import { LaunchTimer, noopClusterMetrics, type ClusterMetrics } from '../metrics/cluster-metrics.js'
 import { Backoff, systemClock, type Clock } from '@agentconnect.md/connection'
 import type { SpawnDriver, SpawnRequest, SpawnedRuntime } from '../acp/spawn-driver.js'
 import type { ShimCapability } from '../shim/protocol.js'
