+++ packages/daemon/test/cluster-metrics.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it, vi } from 'vitest'
 import { randomUUID } from 'node:crypto'
 import { FakeClock } from '@agentconnect.md/connection'
 import { K8sDriver, type ClusterDriverDeps } from '../src/k8s/driver.js'
-import { LaunchTimer, type ClusterMetrics, type LaunchPath, type LaunchStage } from '../src/k8s/cluster-metrics.js'
+import { LaunchTimer, type ClusterMetrics, type LaunchPath, type LaunchStage } from '../src/metrics/cluster-metrics.js'
 import { K8sApiError } from '@agentconnect.md/k8s-client'
 import { GuardedResumeRejectedError, type Sandbox, type SandboxClaim } from '../src/k8s/sandbox-api.js'
 import { fakeGenerations } from './fake-generations.js'
